requirements.txt
tweepy==3.5.0
gspread==0.3.0
oauth2client==1.5.2
pyopenssl
python-memcached
